Major format change
+divided files into suitable sub-folders
+added hashFileEstimater
+updated README
