--- Post.py
@@ -1,20 +1,0 @@
-class Post:
-	url=""
-	timePosted=0
-	phash=""
-	guild=""
-
-	def __init__(self, url, timePosted, phash, guild):
-		self.url = url
-		self.timePosted = timePosted
-		self.phash = phash
-		self.guild = guild
-
-	def __str__(self):
-		return  f'Post({self.url}, {self.timePosted}, {self.phash})'
-
-	def to_dict(p):
-		if isinstance(p, Post):
-			return {"url": p.url, "timePosted": p.timePosted, "phash": p.phash, "guild": p.guild}
-		else:
-			raise TypeError("Unexpected type {0}".format(p.__class__.__name__))
