+++ README.md
@@ -10,9 +10,16 @@ private Hydrus library, it's all up to you.
 
 Usage:
 If you just want to work on tagging your Hydrus files the simplest way is to run HyTagUI.py
+
+  It is highly recommended to run hashFileEstimater.py first. It will rapidly go through your Hydrus database and create two files for you. This is extremely useful if it is your first time using
+  HyTagUI.py or if you, like I did 30 minutes prior to making hashFileEstimater.py, accidentally delete your file containing all of the processed files HyTagUI has already worked on.
+  processedHashes.txt will contain a list of all hashes that have at least one url reference to a site that has tags (Gel, Dan, Sankaku, E621, Pixiv) and is the default for HyTagUI.py. 
+  needWork.txt doesn't actually have much use and can be deleted.
+
+  (Make sure you have a processedHashes.txt file, even if it's blank)
   Launch HyTagUI.py
-  Hit "Initialize" and wait, the more files you have the longer it will take as it is checking all files to see which need tagging. Hydrus is going to be bombarded when this happens and it currently can not be paused during this stage, so don't plan on being able to use Hydrus very much during this time.
-  Once it is done it will let you know you can start. Set the stop limit (default 100) to whatever you want and hit "Run", but know that the base limit for sauceNao is 200. Personally I     recommend 150 as your limit.
+  Hit "Initialize" and wait, the more files you have the longer it will take as it is checking all files to see which need tagging. Hydrus is going to be bombarded when this happens and it currently can not be paused during this stage, so don't plan on being able to use Hydrus very much during this time. At around 55,000 files it takes me up to 15 minutes to initialize. Don't worry, optimization will be coming at some point.
+  Once it is done it will let you know you can start. Set the stop limit (default 100) to whatever you want and hit "Run", but know that the base limit for sauceNao is 200. Personally I recommend 150 as your limit.
   Hit run. It will begin tagging your images using the tags it finds from GelBooru, DanBooru, Yandex, Twitter, and many more. This WILL slow down Hydrus a little bit, but you can pause the program with the big red "Pause" button.
   When it's done, or you have paused it, you can hit "Stop" to close the program. This is more of a formality, if you really want you can just close it anyway. Try to use the "stop" button though as it is safer and less likely to cause any possible issues with the file tracking what hashes have been searched/tagged.
