--- Runner.py
@@ -1,8 +1,0 @@
-import HydrusApi
-import sauceNaoApi
-import sys
-
-
-fileLoc = HydrusApi.getImageById(sys.argv[1])
-tags = sauceNaoApi.getAllTagsFromFile(fileLoc)
-print(tags)
